refactor: un solo helper de actaRecibido
